move "vcpatch_code_table.c" all to "vcpatch_wrapper.c";
