+++ builds/vc2019/HPatchZ.vcxproj
@@ -376,7 +376,6 @@
     <ClCompile Include="..\..\hpatchz.c" />
     <ClCompile Include="..\..\libHDiffPatch\HDiff\private_diff\limit_mem_diff\adler_roll.c" />
     <ClCompile Include="..\..\libHDiffPatch\HPatch\patch.c" />
-    <ClCompile Include="..\..\vcdiff_wrapper\vcpatch_code_table.c" />
     <ClCompile Include="..\..\vcdiff_wrapper\vcpatch_wrapper.c" />
   </ItemGroup>
   <ItemGroup>
